requirements.txt
pytest==7.4.4
pytest-selenium==4.1.0
selenium==4.17.2
webdriver-manager==4.0.1
keyboard==0.13.5
pytest-html==4.1.0
allure-pytest==2.13.5
pytest-xdist==3.6.1
pytest-cov==5.0.0
pytest-bdd==7.1.2
coverage==7.5.4
requests==2.32.3

